lambdas/pkgevents/requirements.txt
boto3==1.17.15
python-dateutil==2.8.1
